Rename wayland-util -> registry
