+++ src/gtk4-layer-shell.c
@@ -1,5 +1,5 @@
 #include "gtk4-layer-shell.h"
-#include "wayland-utils.h"
+#include "registry.h"
 #include "layer-surface.h"
 #include "libwayland-shim.h"
