+++ src/registry.h
